frontend/requirements.txt
streamlit==1.31.0
pandas
requests
python-dotenv